* 'master' of https://github.com/SmallElephant/FEAlgorithm-Swift:
  update test case~
  二叉搜索树完善~
  二叉搜索树~
  二叉树修改~
  栈实现~
  update~
  二叉树的镜像~
  二叉树创建修改~

Conflicts:
	5-Tree/5-Tree.xcodeproj/project.pbxproj
	5-Tree/5-Tree/ConstructBinaryTree.swift
	5-Tree/5-Tree/StringExtension.swift
	5-Tree/5-Tree/TreeNode.swift
	5-Tree/5-Tree/TreeOrder.swift
	5-Tree/5-Tree/main.swift

+++ 5-Tree/5-Tree/StringExtension.swift
@@ -10,15 +10,15 @@ import Foundation
 
 extension String {
     
-    subscript(index:Int)->Character {
+    subscript(index:NSInteger)->Character {
         return self[self.characters.index(self.startIndex, offsetBy: index)]
     }
     
-    subscript(index:Int)->String {
+    subscript(index:NSInteger)->String {
         return String(self[index] as Character)
     }
     
-    subscript(range:Range<Int>)->String {
+    subscript(range:Range<NSInteger>)->String {
         let start = characters.index(startIndex, offsetBy: range.lowerBound)
         let end = characters.index(start, offsetBy: range.upperBound-range.lowerBound)
         return self[Range(start..<end)]
